llm-service/requirements.txt
flask-cors
flask
openai>=1.0.0
python-dotenv

